v1.8.73: implement card emulation via PN532 TgInitAsTarget
New pn532_target.c: puts PN532 in NFC-A target mode using TgInitAsTarget
(0x8C). Presents card's ATQA/SAK/UID[0..2] (3 bytes; PN532 completes to
4-byte UID). Handles READ(0x30) with saved page data, GET_VERSION(0x60),
and NAKs unknown commands. Stop latency <= 3 s.

rfid_manager: add start/stop/is_emulating with FreeRTOS task (mirrors
poll task design). rfid_back_to_menu stops emulation on back navigation.

UI: replaces placeholder with card info panel, Emulate/Stop buttons,
and status label (Waiting / Reader active! / Stopped / Error).

Also fix: rfid_storage_ensure_dirs parent dir bug (v1.8.72 backport
already shipped), gatt_walker.c format-specifier warning.

+++ ESP32C5/main/gatt_walker.c
@@ -765,7 +765,7 @@ static void s_make_timestamp(char *out, size_t len)
     } else {
         /* RTC not set — use uptime in seconds */
         uint32_t secs = (uint32_t)(esp_timer_get_time() / 1000000ULL);
-        snprintf(out, len, "00000000_%06u", secs % 1000000u);
+        snprintf(out, len, "00000000_%06u", (unsigned int)(secs % 1000000u));
     }
 }
